Cleanup: move bmqsys_operationlogger -> bmqu_operationlogger
