+++ src/groups/bmq/bmqsys/package/bmqsys.mem
@@ -1,4 +1,3 @@
 bmqsys_mocktime
-bmqsys_operationlogger
 bmqsys_statmonitor
 bmqsys_statmonitorsnapshotrecorder
