+++ src/groups/bmq/bmqu/bmqu_operationlogger.h
@@ -13,16 +13,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef INCLUDED_BMQSYS_OPERATIONLOGGER
-#define INCLUDED_BMQSYS_OPERATIONLOGGER
+#ifndef INCLUDED_BMQU_OPERATIONLOGGER
+#define INCLUDED_BMQU_OPERATIONLOGGER
 
 //@PURPOSE: Provide utility to measure and log operation time.
 //
 //@CLASSES:
-// bmqsys::OperationLogger
+// bmqu::OperationLogger
 //
 //@DESCRIPTION:
-// This component provides an utility class `bmqsys::OperationLogger` that can
+// This component provides an utility class `bmqu::OperationLogger` that can
 // be used to measure and log time spent to perform operation.
 // There are several benefits of using this class:
 //: o Ability to perform measures and to log them in an uniform style.
@@ -31,7 +31,7 @@
 //: o Simplify classes that measure operation times by removing their internal
 //:   states used to keep track of time and operation name.
 //: o Allow to simply keep track of several async operations at the same time
-//:   by using different instances of `bmqsys::OperationLogger`.
+//:   by using different instances of `bmqu::OperationLogger`.
 //: o Flexibility in usage pattern: operation loggers might be used as fully
 //:   disposable objects and as reusable objects (object pool compatible).
 //
@@ -40,7 +40,7 @@
 // 1. Simple scoped operation logger
 //..
 //  {
-//      bmqsys::OperationLogger opLogger(allocator);
+//      bmqu::OperationLogger opLogger(allocator);
 //      opLogger.start() << "Scoped operation";
 //
 //      doSomething1();
@@ -52,8 +52,8 @@
 //
 // 2. Pass operation logger to another function
 //..
-//  bsl::shared_ptr<bmqsys::OperationLogger> opLogger =
-//      bsl::allocate_shared<bmqsys::OperationLogger>(d_state.d_allocator_p);
+//  bsl::shared_ptr<bmqu::OperationLogger> opLogger =
+//      bsl::allocate_shared<bmqu::OperationLogger>(d_state.d_allocator_p);
 //  opLogger.start() << "Pass operation logger";
 //  // Make sure to `move` so we don't have a copy of this shared pointer in
 //  // this scope for accurate results.
@@ -63,14 +63,14 @@
 // 3. Use pool for operation loggers
 //..
 //  typedef bdlcc::SharedObjectPool<
-//      bmqsys::OperationLogger,
+//      bmqu::OperationLogger,
 //      bdlcc::ObjectPoolFunctors::DefaultCreator,
-//      bdlcc::ObjectPoolFunctors::Clear<bmqsys::OperationLogger> >
+//      bdlcc::ObjectPoolFunctors::Clear<bmqu::OperationLogger> >
 //      OperationLoggerPool;
 //
 //  OperationLoggerPool pool(-1, allocator);
 //  for (int i = 0; i < 100; i++) {
-//      bsl::shared_ptr<bmqsys::OperationLogger> opLogger = pool.getObject();
+//      bsl::shared_ptr<bmqu::OperationLogger> opLogger = pool.getObject();
 //      opLogger.start() << "Dynamic operation " << i;
 //      thread.doSomethingAsync(opLogger);
 //  }
@@ -91,7 +91,7 @@
 #include <bmqu_time.h>
 
 namespace BloombergLP {
-namespace bmqsys {
+namespace bmqu {
 
 // =====================
 // class OperationLogger
@@ -110,7 +110,7 @@ class OperationLogger BSLS_KEYWORD_FINAL {
     bmqu::MemOutStream d_currentOpDescription;
 
     // CLASS-SCOPE CATEGORY
-    BALL_LOG_SET_CLASS_CATEGORY("BMQSYS.OPERATIONLOGGER");
+    BALL_LOG_SET_CLASS_CATEGORY("BMQU.OPERATIONLOGGER");
 
     // NOT IMPLEMENTED
     /// Copy and assignment are disabled to prevent misuse of this class.
