+++ src/groups/bmq/bmqu/package/bmqu.mem
@@ -13,6 +13,7 @@ bmqu_objectplaceholder
 bmqu_objectplaceholder_cpp03
 bmqu_operationchain
 bmqu_operationchain_cpp03
+bmqu_operationlogger
 bmqu_outstreamformatsaver
 bmqu_printutil
 bmqu_samethreadchecker
